% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{back}
\alias{back}
\title{Advance in the game tree}
\usage{
back(game, steps = 1)
}
\arguments{
\item{game}{A game}

\item{steps}{How many steps (half-turns) to go back}
}
\value{
A game
}
\description{
Advance in the game tree
}
