% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{board_color}
\alias{board_color}
\alias{has_insufficient_material}
\alias{has_castling_rights}
\alias{has_kingside_castling_rights}
\alias{has_queenside_castling_rights}
\title{Get information about the current board given a color}
\usage{
has_insufficient_material(game, color)

has_castling_rights(game, color)

has_kingside_castling_rights(game, color)

has_queenside_castling_rights(game, color)
}
\arguments{
\item{game}{A game node}

\item{color}{Color to use (\code{TRUE} is White and \code{FALSE} is Black)}
}
\value{
A boolean
}
\description{
Get information about the current board given a color
}
