% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{board_is}
\alias{board_is}
\alias{is_checkmate}
\alias{is_check}
\alias{is_game_over}
\alias{is_stalemate}
\alias{is_insufficient_material}
\alias{is_seventyfive_moves}
\alias{is_fivefold_repetition}
\alias{is_repetition}
\alias{can_claim_draw}
\alias{can_claim_fifty_moves}
\alias{can_claim_threefold_repetition}
\alias{has_en_passant}
\title{Get information about the current board}
\usage{
is_checkmate(game)

is_check(game)

is_game_over(game)

is_stalemate(game)

is_insufficient_material(game)

is_seventyfive_moves(game)

is_fivefold_repetition(game)

is_repetition(game, count = 3)

can_claim_draw(game)

can_claim_fifty_moves(game)

can_claim_threefold_repetition(game)

has_en_passant(game)
}
\arguments{
\item{game}{A game node}

\item{count}{Number of moves to count for repetition}
}
\value{
A boolean
}
\description{
Get information about the current board
}
