% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{board_move}
\alias{board_move}
\alias{gives_check}
\alias{is_en_passant}
\alias{is_capture}
\alias{is_zeroing}
\alias{is_irreversible}
\alias{is_castling}
\alias{is_kingside_castling}
\alias{is_queenside_castling}
\title{Get information about the current board given a move}
\usage{
gives_check(game, move, notation = c("san", "uci", "xboard"))

is_en_passant(game, move, notation = c("san", "uci", "xboard"))

is_capture(game, move, notation = c("san", "uci", "xboard"))

is_zeroing(game, move, notation = c("san", "uci", "xboard"))

is_irreversible(game, move, notation = c("san", "uci", "xboard"))

is_castling(game, move, notation = c("san", "uci", "xboard"))

is_kingside_castling(game, move, notation = c("san", "uci", "xboard"))

is_queenside_castling(game, move, notation = c("san", "uci", "xboard"))
}
\arguments{
\item{game}{A game node}

\item{move}{Move to consider}

\item{notation}{Notation used for \code{move}}
}
\value{
A boolean
}
\description{
Get information about the current board given a move
}
