% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{board_to_string}
\alias{board_to_string}
\title{Convert a board to either unicode or ASCII string}
\usage{
board_to_string(x, unicode = FALSE, invert_color = FALSE, empty_square = ".")
}
\arguments{
\item{x}{A board}

\item{unicode}{Use unicode characters?}

\item{invert_color}{Invert piece color? Useful for white text on dark
background.}

\item{empty_square}{Character used for empty square}
}
\value{
A string
}
\description{
Convert a board to either unicode or ASCII string
}
