% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{branch}
\alias{branch}
\title{Add next move as one branch of the tree}
\usage{
branch(game, move, notation = c("san", "uci", "xboard"), enter = TRUE)
}
\arguments{
\item{game}{A game node}

\item{move}{Description of a move}

\item{notation}{Notation used for \code{move}}

\item{enter}{Follow move into new branch? Works like \verb{git checkout}}
}
\value{
A game node
}
\description{
Add next move as one branch of the tree
}
