% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\name{fish}
\alias{fish}
\title{Have Stockfish play the next move}
\usage{
fish(game, time = 0.1)
}
\arguments{
\item{game}{A game node}

\item{time}{Time limit (in seconds) allocated to Stockfish}
}
\value{
A game node
}
\description{
Have Stockfish play the next move
}
