% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\name{fish_configure}
\alias{fish_configure}
\title{Configure and start Stockfish engine}
\usage{
fish_configure(path = NULL, options = list())
}
\arguments{
\item{path}{Path to Stockfish executable (defaults to bundled version)}

\item{options}{A list of options passed on to Stockfish}
}
\description{
Configure and start Stockfish engine
}
