% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{forward}
\alias{forward}
\title{Advance in the game tree, playing next move from current branch}
\usage{
forward(game, steps = 1)
}
\arguments{
\item{game}{A game node}

\item{steps}{How many steps (half-turns) to advance}
}
\value{
A game node
}
\description{
Advance in the game tree, playing next move from current branch
}
