% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{halfmove_clock}
\alias{halfmove_clock}
\title{Get number of half-moves since the last capture or pawn move}
\usage{
halfmove_clock(game)
}
\arguments{
\item{game}{A game node}
}
\value{
An integer
}
\description{
Get number of half-moves since the last capture or pawn move
}
