% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{line}
\alias{line}
\title{Branch game with next move}
\usage{
line(game, moves, notation = c("san", "uci", "xboard"), enter = FALSE)
}
\arguments{
\item{game}{A game node}

\item{moves}{Vector of one or more description of moves}

\item{notation}{Notation used for \code{moves}}

\item{enter}{Follow new branch to the end? Works like \verb{git checkout}}
}
\value{
A game node
}
\description{
Branch game with next move
}
