% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{move}
\alias{move}
\title{Make moves and create variations}
\usage{
move(game, ..., notation = c("san", "uci", "xboard"))
}
\arguments{
\item{game}{A game node}

\item{...}{Sequence of moves (lists are converted to a variation the same
way parentheses work for PGN)}

\item{notation}{Notation used for moves}
}
\value{
A game node
}
\description{
Make moves and create variations
}
