% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{move_}
\alias{move_}
\title{Make moves and create variations}
\usage{
move_(game, moves, notation = c("san", "uci", "xboard"))
}
\arguments{
\item{game}{A game node}

\item{moves}{List of moves}

\item{notation}{Notation used for moves}
}
\value{
A game node
}
\description{
Make moves and create variations
}
