% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nag.R
\name{nag}
\alias{nag}
\title{Parse Numeric Annotation Glyph (NAG) of a move}
\usage{
nag(game)
}
\arguments{
\item{game}{A game node}
}
\value{
A string
}
\description{
Parse Numeric Annotation Glyph (NAG) of a move
}
