% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{parse_move}
\alias{parse_move}
\title{Parse move in context}
\usage{
parse_move(game, moves, notation = c("san", "uci", "xboard"))
}
\arguments{
\item{game}{A game node}

\item{moves}{A move string}

\item{notation}{Notation used for \code{move}}
}
\value{
A move object
}
\description{
Parse move in context
}
