% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{play}
\alias{play}
\title{Move a piece on the board}
\usage{
play(game, moves, notation = c("san", "uci", "xboard"))
}
\arguments{
\item{game}{A game node}

\item{moves}{Vector of one or more description of moves}

\item{notation}{Notation used for \code{moves}}
}
\value{
A game node
}
\description{
Move a piece on the board
}
