% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.chess.pgn.GameNode}
\alias{plot.chess.pgn.GameNode}
\title{Plot game node}
\usage{
\method{plot}{chess.pgn.GameNode}(x, ...)
}
\arguments{
\item{x}{A game node}

\item{...}{Not used}
}
\description{
Plot game node
}
