% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.chess.Board}
\alias{print.chess.Board}
\title{Print board}
\usage{
\method{print}{chess.Board}(x, unicode = FALSE, invert_color = FALSE, empty_square = ".", ...)
}
\arguments{
\item{x}{A game board}

\item{unicode}{Use unicode characters?}

\item{invert_color}{Invert piece color? Useful for white text on dark
background.}

\item{empty_square}{Character used for empty square}

\item{...}{Not used}
}
\description{
Print board
}
