% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{print.chess.pgn.GameNode}
\alias{print.chess.pgn.GameNode}
\title{Print game board}
\usage{
\method{print}{chess.pgn.GameNode}(x, unicode = FALSE, invert_color = FALSE, empty_square = ".", ...)
}
\arguments{
\item{x}{A game board}

\item{unicode}{Use unicode characters?}

\item{invert_color}{Invert piece color? Useful for white text on dark
background.}

\item{empty_square}{Character used for empty square}

\item{...}{Not used}
}
\description{
Print game board
}
