% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.chess.pgn.Variations}
\alias{print.chess.pgn.Variations}
\title{Print a list of variations}
\usage{
\method{print}{chess.pgn.Variations}(x, unicode = FALSE, invert_color = FALSE, empty_square = ".", ...)
}
\arguments{
\item{x}{A game node}

\item{unicode}{Use unicode characters?}

\item{invert_color}{Invert piece color? Useful for white text on dark
background}

\item{empty_square}{Character used for empty square}

\item{...}{Not used}
}
\description{
Print a list of variations
}
