% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn.R
\name{read_game}
\alias{read_game}
\title{Read a game from a PGN}
\usage{
read_game(file, n_max = Inf)
}
\arguments{
\item{file}{File or connection to read from}

\item{n_max}{Maximum number of games to read}
}
\value{
A game node
}
\description{
Read a game from a PGN
}
