% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{result}
\alias{result}
\title{Get result of the game ("*" if it hasn't ended)}
\usage{
result(game)
}
\arguments{
\item{game}{Any node of a game}
}
\value{
A string
}
\description{
Get result of the game ("*" if it hasn't ended)
}
