% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\name{stockfish_download}
\alias{stockfish_download}
\title{Download Stockfish executable according to platform}
\usage{
stockfish_download(path = "~/.stockfish")
}
\arguments{
\item{path}{Path where to save executable}
}
\value{
Path to executable
}
\description{
Download Stockfish executable according to platform
}
