% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{turn}
\alias{turn}
\title{Get whose turn it is}
\usage{
turn(game)
}
\arguments{
\item{game}{A game node}
}
\value{
A boolean (\code{TRUE} is White and \code{FALSE} is Black)
}
\description{
Get whose turn it is
}
