% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{undo}
\alias{undo}
\title{Undo last move}
\usage{
undo(board, steps = 1)
}
\arguments{
\item{board}{A board}

\item{steps}{How many moves (half-turns) to revert}
}
\value{
A board
}
\description{
Undo last move
}
