% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{variation}
\alias{variation}
\title{Follow variation of a move}
\usage{
variation(game, id = 1)
}
\arguments{
\item{game}{A game node}

\item{id}{Index of variation (1 is the mainline)}
}
\value{
A game node
}
\description{
Follow variation of a move
}
