% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{variations}
\alias{variations}
\title{Get all variations for next move (the children of current node)}
\usage{
variations(game)
}
\arguments{
\item{game}{A game node}
}
\value{
A list of games nodes
}
\description{
Get all variations for next move (the children of current node)
}
