% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn.R
\name{write_game}
\alias{write_game}
\title{Save a game as an PGN}
\usage{
write_game(x, file)
}
\arguments{
\item{x}{Any node of a game}

\item{file}{File or connection to write to}
}
\description{
Save a game as an PGN
}
