% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{back}
\alias{back}
\title{Go back in the game tree, reverting the last move from current branch}
\usage{
back(game, steps = 1)
}
\arguments{
\item{game}{A game node}

\item{steps}{How many steps (half-turns) to go back}
}
\value{
A game node
}
\description{
Go back in the game tree, reverting the last move from current branch
}
