% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{install_chess}
\alias{install_chess}
\title{Install python-chess}
\usage{
install_chess(method = "auto", conda = "auto", ...)
}
\arguments{
\item{method}{Installation method}

\item{conda}{The path to a conda executable}

\item{...}{Other arguments passed on to \code{\link[reticulate:py_install]{reticulate::py_install()}}}
}
\description{
Install python-chess
}
