% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{move}
\alias{move}
\title{Move a piece on the board}
\usage{
move(game, move, notation = c("san", "uci", "xboard"))
}
\arguments{
\item{game}{A game node}

\item{move}{Description of a move}

\item{notation}{Notation used for \code{move}}
}
\value{
A game node
}
\description{
Move a piece on the board
}
