% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{move}
\alias{move}
\title{Create subtree of game with next move}
\usage{
move(game, moves, notation = c("san", "uci", "xboard"))
}
\arguments{
\item{game}{A game node}

\item{moves}{A list or vector of moves (each sublist will be converted to a
variation the same way parentheses work for PGN)}

\item{notation}{Notation used for \code{moves}}
}
\value{
A game node
}
\description{
Create subtree of game with next move
}
