% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{move}
\alias{move}
\title{Move a piece on the board}
\usage{
move(board, move, notation = c("san", "uci", "xboard"))
}
\arguments{
\item{board}{A board}

\item{move}{Description of a move}

\item{notation}{Notation used for \code{move}}
}
\value{
A board
}
\description{
Move a piece on the board
}
