% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\name{stockfish_configure}
\alias{stockfish_configure}
\title{Configure and start Stockfish engine}
\usage{
stockfish_configure(options = list())
}
\arguments{
\item{options}{A list of options passed on to Stockfish}
}
\value{
\code{TRUE} if startup is successful
}
\description{
Configure and start Stockfish engine
}
