% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{write_svg}
\alias{write_svg}
\title{Save an SVG with rendering of the board}
\usage{
write_svg(x, file)
}
\arguments{
\item{x}{A game node}

\item{file}{File or connection to write to}
}
\description{
Save an SVG with rendering of the board
}
